+++ src/fetchDescription.js
@@ -3,6 +3,7 @@ const { error } = require('@actions/core');
 const util = require('./util')
 module.exports = async({authToken,jiraApiUrl}) => {
     try{
+        console.log("Called fetch description");
     const response = await fetch(jiraApiUrl,{
         headers:{ 
             Authorization: `Basic ${authToken}` }
